+++ HISTORY.md
@@ -2,14 +2,19 @@
 
 ---
 
+## 0.4.0 / 2015-12-29
+
+- Add prop `popupVisible`.
+- `onVisibleChange` => `onPopupVisibleChange`.
+
 ## 0.3.0 / 2015-12-28
 
-- Fix value and defaultValue
-- Add Test Case
+- Fix value and defaultValue.
+- Add Test Cases.
 
 ## 0.2.0 / 2015-12-25
 
-- Add `expandTrigger`.
+- Add prop `expandTrigger`.
 
 ## 0.1.0 / 2015-12-25